requirements.txt
flask  
numpy 
tensorflow
keras 
mysql.connector
datetime
cv2
random
shutil
os
pandas 
matplotlib.pyplot 
time

    